requirements.txt
streamlit
pandas
matplotlib
mysql-connector-python
datetime
langgraph
langchain
langchain-aws
boto3
openpyxl